+++ src/PhpDoc/LattePhpDoc.php
@@ -4,24 +4,44 @@ declare(strict_types=1);
 
 namespace Efabrica\PHPStanLatte\PhpDoc;
 
+use PHPStan\ShouldNotHappenException;
+
 final class LattePhpDoc
 {
     private bool $isIgnored;
 
-    public function __construct(bool $isIgnored = false)
+    /** @var array<string>|null  */
+    private ?array $templatePaths = null;
+
+    private ?self $parent = null;
+
+    /**
+     * @param array<string>|null $templatePaths
+     */
+    public function __construct(bool $isIgnored = false, ?array $templatePaths = null)
     {
         $this->isIgnored = $isIgnored;
+        $this->templatePaths = $templatePaths;
     }
 
     public function isIgnored(): bool
     {
-        return $this->isIgnored;
+        return $this->isIgnored || ($this->parent !== null && $this->parent->isIgnored());
     }
 
-    public function merge(self $inherit): self
+    /**
+     * @return array<string>
+     */
+    public function getTemplatePaths(): ?array
     {
-        return new self(
-            $inherit->isIgnored() || $this->isIgnored
-        );
+        return $this->templatePaths;
+    }
+
+    public function setParent(self $parent): void
+    {
+        if ($this->parent !== null) {
+            throw new ShouldNotHappenException('Cannot change LattePhpDoc parent');
+        }
+        $this->parent = $parent;
     }
 }