+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Annotation `@phpstan-latte-ignore` can be used to ignore render calls, variable assignments, component creation, whole methods or classes.
+- Annotation `@phpstan-latte-template` can be used to specify what template is used to render.
 
 ## [0.2.0] - 2022-12-06
 ### Changed